etc: inputsnowball index, folder rename
