+++ front/src/components/InputSnowball/InputSnowball.tsx
