+++ front/src/components/InputSnowball/index.tsx
