+++ front/src/components/index.tsx
@@ -1,7 +1,7 @@
 export { Song } from './Song';
 export { SnowGlobeCanvas } from './SnowGlobeCanvas';
 export { Button, StepButton, PostButton, MakeButton } from './Button';
-export { InputSnowball } from './Input';
+export { InputSnowball } from './InputSnowball';
 export { HeaderText } from './HeaderText';
 export { Prev } from './Prev';
 export { Msg } from './Msg';
